Allow admin requests only within the admin backend

+++ src/Aimeos/Shop/Controller/JqadmController.php
@@ -125,6 +125,14 @@ class JqadmController extends AdminController
 			$this->authorize( 'admin', [JqadmController::class, config( 'shop.roles', ['admin', 'editor'] )] );
 		}
 
+		// allow requests only if they are from within the admin backend
+		$cnt = count( explode( '/', request()->route()->getPrefix() ) ) - 1;
+		$base = request()->root() . '/' . join( '/', array_slice( request()->segments(), 0, $cnt ) );
+
+		if( strncmp( url()->previous(), $base, strlen( $base ) ) ) {
+			abort( 403 );
+		}
+
 		$cntl = $this->createAdmin();
 
 		if( ( $html = $cntl->delete() ) == '' ) {
@@ -186,6 +194,14 @@ class JqadmController extends AdminController
 	{
 		if( config( 'shop.authorize', true ) ) {
 			$this->authorize( 'admin', [JqadmController::class, config( 'shop.roles', ['admin', 'editor'] )] );
+		}
+
+		// allow requests only if they are from within the admin backend
+		$cnt = count( explode( '/', request()->route()->getPrefix() ) ) - 1;
+		$base = request()->root() . '/' . join( '/', array_slice( request()->segments(), 0, $cnt ) );
+
+		if( strncmp( url()->previous(), $base, strlen( $base ) ) ) {
+			abort( 403 );
 		}
 
 		$cntl = $this->createAdmin();
